requirements.txt
PyMuPDF
huaweicloudsdkcore==3.1.123
huaweicloudsdkocr==3.1.123
loguru==0.7.2
openpyxl==3.1.5
pandas==2.2.3
pocode==0.0.3
pofile==0.1.5
poprogress==0.0.2
PyPDF2==3.0.1
PySimpleGUI==5.0.7
Requests==2.32.3
setuptools==69.0.2
tencentcloud_sdk_python==3.0.1271
toml==0.10.2
translate==3.6.1
wftools==0.0.9
